feat: add queued actions count badge
